+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Application
 - Add `EnsureApsUserHaveOpenId` to make sure `OpenId` has data. 
 - Disable `EnsureApsUserHaveOpenId`, not working `OpenId` is null after `Refresh`.
-- Update `OpenFile` in `Log`.
+- Update `OpenFile` in `Log` with `UseShellExecute=true`.
 - Update `net7.0-windows` for Revit 2025.
 - Update `csproj`.
 - Add `ApplicationPreviewUtils` to enable know if Revit application is preview.
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Select resource console zip for NetFramework and NetCore.
 - Update `ResourceConsoleUtils`
 - Update `RevitTestConsole` to remove `+` in FileInfo.
+- Add `FileVersionInfoUtils` to remove `+` in FileInfo.
 ### Warning
 - Ignore `Newtonsoft.Json` warning in `Application` project.